+++ rekono/installation/tools.py
@@ -4,8 +4,7 @@ import subprocess
 import sys
 
 from git import Repo
-
-from rekono.config import GITTOOLS_DIR, LOG4J_SCANNER_DIR
+from rekono.config import GITTOOLS_DIR, LOG4J_SCAN_DIR, SPRING4SHELL_SCAN_DIR
 from rekono.utils.linux.apt import apt_install
 
 
@@ -15,7 +14,8 @@ def install_tools() -> None:
         [
             'nmap', 'dirsearch', 'theharvester', 'nikto', 'sslscan', 'sslyze',
             'cmseek', 'zaproxy', 'exploitdb', 'metasploit-framework',
-            'emailharvester', 'joomscan', 'gitleaks', 'smbmap'
+            'emailharvester', 'joomscan', 'gitleaks', 'smbmap', 'nuclei',
+            'gobuster'
         ],
         required=False
     )
@@ -23,18 +23,22 @@ def install_tools() -> None:
         'sudo', sys.executable, '-m', 'pip', 'install', '-q', 'emailfinder', 'ssh-audit'
     ])
     for git_repository, directory in [
-        ('https://github.com/cisagov/log4j-scanner.git', LOG4J_SCANNER_DIR),
+        ('https://github.com/fullhunt/log4j-scan', LOG4J_SCAN_DIR),
+        ('https://github.com/fullhunt/spring4shell-scan.git', SPRING4SHELL_SCAN_DIR),
         ('https://github.com/internetwache/GitTools.git', GITTOOLS_DIR),
     ]:
         if not os.path.isdir(directory):
             subprocess.run(['sudo', 'mkdir', directory])                        # Create tool directory
             subprocess.run(['sudo', 'chmod', '-R', '777', directory])           # Change tool directory permissions
             Repo.clone_from(git_repository, directory)
-    log4j_scanner_req = os.path.join(LOG4J_SCANNER_DIR, 'log4-scanner', 'requirements.txt')
-    if os.path.isfile(log4j_scanner_req):
-        subprocess.run([                                                        # Install dependencies for Log4J Scanner
-            'sudo', sys.executable, '-m', 'pip', 'install', '-q', '-r', log4j_scanner_req
-        ])
+    for requirements in [
+        os.path.join(LOG4J_SCAN_DIR, 'requirements.txt'),
+        os.path.join(SPRING4SHELL_SCAN_DIR, 'requirements.txt'),
+    ]:
+        if os.path.isfile(requirements):
+            subprocess.run([                                                    # Install Python dependencies
+                'sudo', sys.executable, '-m', 'pip', 'install', '-q', '-r', requirements
+            ])
 
 
 def configure_tools() -> None:
@@ -47,7 +51,4 @@ def configure_tools() -> None:
 
 def install_resources() -> None:
     '''Install all resources used by Rekono.'''
-    apt_install(['wordlists', 'seclists', 'dirb'])
-    rockyou = '/usr/share/wordlists/rockyou.txt'
-    if os.path.isfile(f'{rockyou}.gz') and not os.path.isfile(rockyou):
-        subprocess.run(['sudo', 'gzip', '-d', f'{rockyou}.gz'], capture_output=True)    # Decompress rockyou wordlist
+    apt_install(['seclists', 'dirb'])